refactor: Crear custom hook para detalles de Completar Oracion. #167
